requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.90.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.4.0
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.23.1
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
